refactor: move get_connection_id to its module inside UDP module
